requirements.txt
attrs==19.3.0
Automat==0.8.0
bcrypt==3.1.7
cffi==1.13.2
constantly==15.1.0
cryptography==2.8
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.19
PyHamcrest==1.9.0
pyOpenSSL==19.1.0
service-identity==18.1.0
six==1.13.0
Twisted==19.10.0
zope.interface==4.7.1
